+++ benchmark/equal-circular.js
@@ -10,7 +10,8 @@ const {replacer, ignoreKeys, cloner} = require('./fixtures/utils');
 
 const { ElSegundo } = require('..');
 
-const deepdiff = require('deep-diff');
+// const deepdiff = require('deep-diff');
+// const fastDeepEqual = require('fast-deep-equal');
 
 fixture.sub.self = fixture;
 const fixtureClone = clone(fixture, cloner);
@@ -59,5 +60,9 @@ test('circular objects unchanged, dirty check returns false', function (t) {
     s.bench('lodash.isMatchWith', () => {
       dirtyResult = !isMatchWith(subject, fixtureClone, ignoreKeys);
     });
+
+    /* s.bench('fast-deep-equal', () => {
+      dirtyResult = !fastDeepEqual(subject, fixtureClone);
+    }); */
   });
 });