+++ benchmark/equal-ignore.js
@@ -10,8 +10,6 @@ const {replacer, ignoreKeys, cloner} = require('./fixtures/utils');
 
 const { ElSegundo } = require('..');
 
-const deepdiff = require('deep-diff');
-
 const fixtureString = JSON.stringify(fixture, replacer);
 const fixtureClone = clone(fixture, cloner);
 
@@ -32,8 +30,6 @@ test('ignored value changes, dirty check returns false', function (t) {
       dirtyResult = null;
     });
 
-
-    
     s.burn('Theoretical max', () => {
       dirtyResult = 'a' !== 'a';
     });
@@ -45,12 +41,6 @@ test('ignored value changes, dirty check returns false', function (t) {
     s.bench('El Segundo', () => {
       dirtyResult = isDirty.check(subject);
     });
-
-    /*
-    s.bench('deep-diff', () => {
-      dirtyResult = typeof deepdiff(fixtureClone, b) === 'object';
-    });
-    */
 
     s.bench('lodash.isEqualWith', () => {
       dirtyResult = !isEqualWith(subject, fixtureClone, ignoreKeys);