+++ benchmark/not-equal-circular-change.js
@@ -6,11 +6,9 @@ const isMatchWith = require('lodash.ismatchwith');
 const clone = require('lodash.clonedeepwith');
 
 const fixture = clone(require('./fixtures/fixture'));
-const {replacer, ignoreKeys, cloner} = require('./fixtures/utils');
+const { ignoreKeys, cloner } = require('./fixtures/utils');
 
 const { ElSegundo } = require('..');
-
-const deepdiff = require('deep-diff');
 
 fixture.sub.self = fixture;
 const fixtureClone = clone(fixture, cloner);
@@ -36,22 +34,9 @@ test('circular reference changed, dirty check returns true', function (t) {
       dirtyResult = 'a' !== 'b';
     });
 
-    /*
-    Not supported
-    s.bench('JSON.stringify', () => {
-      dirtyResult = fixtureString !== JSON.stringify(b);
-    }); */
-
     s.bench('El Segundo', () => {
       dirtyResult = isDirty.check(subject);
     });
-
-    /*
-    Not supported
-    s.bench('deep-diff', () => {
-      dirtyResult = typeof deepdiff(fixtureClone, b) === 'object';
-    });
-    */
 
     s.bench('lodash.isEqualWith', () => {
       dirtyResult = !isEqualWith(subject, fixtureClone, ignoreKeys);