+++ benchmark/not-equal-key-removed.js
@@ -6,11 +6,12 @@ const isMatchWith = require('lodash.ismatchwith');
 const clone = require('lodash.clonedeepwith');
 
 const fixture = require('./fixtures/fixture');
-const {replacer, ignoreKeys, cloner} = require('./fixtures/utils');
+const { ignoreKeys, cloner } = require('./fixtures/utils');
 
 const { ElSegundo } = require('..');
 
 const deepdiff = require('deep-diff');
+const fastDeepEqual = require('fast-deep-equal');
 
 const fixtureString = JSON.stringify(fixture);
 const fixtureClone = clone(fixture, cloner);
@@ -54,6 +55,10 @@ test('key removed, dirty check returns true', function (t) {
 
     s.bench('lodash.isMatchWith', () => {
       dirtyResult = !isMatchWith(subject, fixtureClone, ignoreKeys);
+    });
+
+    s.bench('fast-deep-equal', () => {
+      dirtyResult = !fastDeepEqual(subject, fixtureClone);
     });
   });
 });